Split Math away from FlatAffineConstraints to PresburgerBasicSet
