+++ mlir/lib/Analysis/Presburger/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_mlir_library(MLIRPresburger
   Simplex.cpp
   Matrix.cpp
+  PresburgerBasicSet.cpp
 
   DEPENDS
   MLIRBuiltinLocationAttributesIncGen
